refactor: separate GitHub API and GitHub Repository Info
